requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flask-wtf==1.0.1